fixed for GLASSFISH-18666: Many warning messages in autodeploy and autoundeploy.
git-svn-id: https://svn.java.net/svn/glassfish~svn/trunk/main@55810 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5

+++ nucleus/deployment/autodeploy/src/main/java/org/glassfish/deployment/autodeploy/AutoOperation.java
@@ -213,9 +213,11 @@ public abstract class AutoOperation {
         try {
             for (String suffix : autoDeployFileSuffixes) {
                 final File suffixedFile = getSuffixedFile(f, suffix);
-                if ( ! suffixedFile.delete()) {
-                    sLogger.log(Level.WARNING, "enterprise.deployment.deleteFailed",
-                            suffixedFile.getAbsolutePath());
+                if(suffixedFile.exists()){
+                    if ( ! suffixedFile.delete()) {
+                        sLogger.log(Level.WARNING, "enterprise.deployment.deleteFailed",
+                                suffixedFile.getAbsolutePath());
+                    }
                 }
             }
         } catch (Exception e) {
